fix: 修复编译错误 - 第一轮 (f64 Eq, unsafe, Pid类型)
- 移除f64字段的Eq约束，改为PartialEq (3个地方)
- 移除AtomicU64的Clone约束
- 将Pid从usize改为i32以支持-1值
- 添加缺失的unsafe块 (copyout, kfree)
- 修复NetworkError的Eq约束

已修复: 7个bug
剩余: 154个编译错误 (从141增加)

+++ kernel/src/posix/mod.rs
@@ -10,7 +10,7 @@ use alloc::string::ToString;
 // ============================================================================
 
 /// Process ID type
-pub type Pid = usize;
+pub type Pid = i32;
 
 /// User ID type  
 pub type Uid = u32;
